outpost/event: fix websocket connection with AUTHENTIK_INSECURE=true
