+++ src/outpost/proxy/upstream.rs
@@ -1,18 +1,12 @@
 //! HTTP client used to forward requests to upstream application servers.
 
-use std::sync::Arc;
-
+use ak_common::tls;
 use axum::body::Body;
 use eyre::Result;
 use hyper_rustls::{HttpsConnector, HttpsConnectorBuilder};
 use hyper_util::{
     client::legacy::{Client, connect::HttpConnector},
     rt::TokioExecutor,
-};
-use rustls::{
-    ClientConfig, DigitallySignedStruct, Error as RustlsError, SignatureScheme,
-    client::danger::{HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier},
-    pki_types::{CertificateDer, ServerName, UnixTime},
 };
 
 /// Client forwarding to upstream servers (HTTP or HTTPS, with optional upgrades).
@@ -23,7 +17,7 @@ pub(super) type UpstreamClient = Client<HttpsConnector<HttpConnector>, Body>;
 pub(super) fn build_client(insecure: bool) -> Result<UpstreamClient> {
     let builder = HttpsConnectorBuilder::new();
     let connector = if insecure {
-        builder.with_tls_config(insecure_tls_config())
+        builder.with_tls_config(tls::client::insecure_config())
     } else {
         builder.with_native_roots()?
     }
@@ -36,56 +30,4 @@ pub(super) fn build_client(insecure: bool) -> Result<UpstreamClient> {
         .set_host(false)
         .http1_title_case_headers(true)
         .build(connector))
-}
-
-fn insecure_tls_config() -> ClientConfig {
-    ClientConfig::builder()
-        .dangerous()
-        .with_custom_certificate_verifier(Arc::new(NoCertificateVerification))
-        .with_no_client_auth()
-}
-
-/// Certificate verifier that accepts any upstream certificate.
-#[derive(Debug)]
-struct NoCertificateVerification;
-
-#[expect(
-    clippy::missing_trait_methods,
-    reason = "the trait's defaulted methods are appropriate for an accept-all verifier"
-)]
-impl ServerCertVerifier for NoCertificateVerification {
-    fn verify_server_cert(
-        &self,
-        _end_entity: &CertificateDer<'_>,
-        _intermediates: &[CertificateDer<'_>],
-        _server_name: &ServerName<'_>,
-        _ocsp_response: &[u8],
-        _now: UnixTime,
-    ) -> Result<ServerCertVerified, RustlsError> {
-        Ok(ServerCertVerified::assertion())
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, RustlsError> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, RustlsError> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
-        rustls::crypto::aws_lc_rs::default_provider()
-            .signature_verification_algorithms
-            .supported_schemes()
-    }
 }
